# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sports World EPS",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/api_eps/public/"
        }
    ],
    "paths": {
...
        "/api/crm2/v1/clubs": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "operationId": "198eff06348ca4eca09ec632cc560ad8",
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "Autorización inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/crm2/v1/disciplines": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "operationId": "6a9da0e35ff3f80a1908d85b80469f4c",
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "Autorización inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/crm2/v1/profileApp/{mail}": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "operationId": "9f2ad28ae137b1cebdd67ee22ea4f6c6",
                "parameters": [
                    {
                        "name": "mail",
                        "in": "path",
                        "schema": {
                            "type": "string",
                            "default": "correo@sports"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "Autorización inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/crm2/v1/updateProfile": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "operationId": "cbe3643db1c9a00cd6f78f4696d2e227",
                "requestBody": {
                    "$ref": "#/components/requestBodies/updateProfileRequest"
                },
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "Autorización inválida"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "requestBodies": {
...
            "updateProfileRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "clientId": {
                                    "type": "integer",
                                    "example": ""
                                },
                                "instructorId": {
                                    "type": "integer",
                                    "example": ""
                                },
                                "productoId": {
                                    "type": "integer",
                                    "example": ""
                                },
                                "clubId": {
                                    "type": "integer",
                                    "example": ""
                                },
                                "responsableId": {
                                    "type": "integer",
                                    "example": ""
                                },
                                "precioId": {
                                    "type": "integer",
                                    "example": ""
                                },
                                "unidades": {
                                    "type": "integer",
                                    "example": ""
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "secret-key",
                "in": "header"
            }
        }
    }
}